// src/core/rate/domain/models/create_rate_model.ts
import type { Option } from "@/src/common/utils/types";
import type { ClientType } from "@/src/core/app/enums/client_type";
import type { PriceType } from "@/src/core/app/enums/price_type";
import type { EnergyTypes } from "@/src/core/app/enums/energy_types";

export class CreateRateModel {
  marketer: Option;
  name: string;
  energyType: EnergyTypes;
  priceType: PriceType;
  clientTypes: ClientType[];
  rateType: Option;
  minPower?: number;
  maxPower?: number;
  permanency: boolean;
  length: string;
  isFullRenewable?: boolean;
  compensationSurplus?: boolean;
  compensationSurplusValue?: number;
  minConsumption?: number;
  maxConsumption?: number;
  energyPrice1?: number;
  energyPrice2?: number;
  energyPrice3?: number;
  energyPrice4?: number;
  energyPrice5?: number;
  energyPrice6?: number;
  powerPrice1?: number;
  powerPrice2?: number;
  powerPrice3?: number;
  powerPrice4?: number;
  powerPrice5?: number;
  powerPrice6?: number;
  fixedTermPrice?: number;

  constructor(rate: {
    marketer: Option;
    name: string;
    energyType: EnergyTypes;
    priceType: PriceType;
    clientTypes: ClientType[];
    rateType: Option;
    permanency: boolean;
    length: string;
    isFullRenewable?: boolean;
    compensationSurplus?: boolean;
    compensationSurplusValue?: number;
    minPower?: number;
    maxPower?: number;
    minConsumption?: number;
    maxConsumption?: number;
    energyPrice1?: number;
    energyPrice2?: number;
    energyPrice3?: number;
    energyPrice4?: number;
    energyPrice5?: number;
    energyPrice6?: number;
    powerPrice1?: number;
    powerPrice2?: number;
    powerPrice3?: number;
    powerPrice4?: number;
    powerPrice5?: number;
    powerPrice6?: number;
    fixedTermPrice?: number;
  }) {
    this.marketer = rate.marketer;
    this.name = rate.name;
    this.energyType = rate.energyType;
    this.priceType = rate.priceType;
    this.clientTypes = rate.clientTypes;
    this.rateType = rate.rateType;
    this.permanency = rate.permanency;
    this.length = rate.length;
    this.isFullRenewable = rate.isFullRenewable;
    this.compensationSurplus = rate.compensationSurplus;
    this.compensationSurplusValue = rate.compensationSurplusValue;
    this.minPower = rate.minPower;
    this.maxPower = rate.maxPower;
    this.minConsumption = rate.minConsumption;
    this.maxConsumption = rate.maxConsumption;
    this.energyPrice1 = rate.energyPrice1;
    this.energyPrice2 = rate.energyPrice2;
    this.energyPrice3 = rate.energyPrice3;
    this.energyPrice4 = rate.energyPrice4;
    this.energyPrice5 = rate.energyPrice5;
    this.energyPrice6 = rate.energyPrice6;
    this.powerPrice1 = rate.powerPrice1;
    this.powerPrice2 = rate.powerPrice2;
    this.powerPrice3 = rate.powerPrice3;
    this.powerPrice4 = rate.powerPrice4;
    this.powerPrice5 = rate.powerPrice5;
    this.powerPrice6 = rate.powerPrice6;
    this.fixedTermPrice = rate.fixedTermPrice;
  }
}